+++ packages/core/web/src/user-accounts/views/GeneralSettings.vue
@@ -1,0 +1,5 @@
+<script lang="ts" setup></script>
+
+<template>Test</template>
+
+<style scoped></style>