+++ packages/libs/ui/src/components/forms/LyCheckbox.stories.ts
@@ -28,7 +28,6 @@ const argTypes = {
   id: { table: { disable: true } },
   name: { table: { disable: true } },
   checked: { table: { disable: true } },
-  required: { table: { disable: true } },
   value: { table: { disable: true } },
   property: { table: { disable: true } },
   error: { table: { disable: true } },