+++ packages/libs/ui/src/components/forms/LyEditableText.stories.ts
@@ -26,7 +26,6 @@ const argTypes = {
   required: { controls: 'boolean' },
   id: { table: { disable: true } },
   name: { table: { disable: true } },
-  required: { table: { disable: true } },
   value: { table: { disable: true } },
   property: { table: { disable: true } },
   error: { table: { disable: true } },